Implemented the eye height for zoans, changed the doriki variable types to longs, commands now support selectors and improved the despawn logic of poison blocks

+++ java/xyz/pixelatedw/MineMineNoMi3/api/telemetry/WyTelemetry.java
@@ -25,7 +25,7 @@ public class WyTelemetry
 			urlConnection = "http://stats.pixelatedw.xyz/globalstats.php";
 	}
 	
-	public static void addStat(final String statName, final int value) 
+	public static void addStat(final String statName, final long value) 
 	{
         Thread newThread = new Thread()
         {
